<!--- Provide a general summary of your changes in the Title above -->

#### Description
<!--- Describe your changes in detail -->
- Only add OnEndOfDay ScheduledEvent if the algorithm implements the
method. **Improves performance in the algorithm thread**. Adding unit tests
- Avoid creating a new baseData instance at
`SubscriptionDataSourceReader`
- Adding static `FineFundamental` instance since creating new ones is
expensive

Performance tests
```PR```
39.67 seconds at 8k data points per second. Processing total of 330,502 data points.
42.38 seconds at 8k data points per second. Processing total of 330,502 data points.
39.26 seconds at 8k data points per second. Processing total of 330,502 data points.

```PR + OnEndOfDay(Symbol) method```
47.81 seconds at 7k data points per second. Processing total of 330,502 data points.
49.88 seconds at 7k data points per second. Processing total of 330,502 data points.
49.62 seconds at 7k data points per second. Processing total of 330,502 data points.

```MASTER + OnEndOfDay(Symbol) method```
49.02 seconds at 7k data points per second. Processing total of 330,502 data points.
46.49 seconds at 7k data points per second. Processing total of 330,502 data points.
45.73 seconds at 7k data points per second. Processing total of 330,502 data points.

- Algorithm used, all daily data reading `SPY`
```
public class EmptyMinute400EquityBenchmark : QCAlgorithm
{
    public override void Initialize()
    {
        SetStartDate(2015, 1, 1);
        SetEndDate(2016, 6, 1);
        foreach (var symbolToAdd in Symbols.Equity.All.Take(100))
        {
            var symbol = AddSecurity(SecurityType.Equity, symbolToAdd, Resolution.Daily, false).Symbol;
            AddData<TiingoNews>(symbol);
        }
    }
    private long _tiingoCount;
    public override void OnData(Slice slice)
    {
        _tiingoCount += slice.Get<TiingoNews>().Count;
    }
    public override void OnEndOfAlgorithm()
    {
        Log($"TiingoCount {_tiingoCount}");
    }
}
```
#### Related Issue
<!--- This project only accepts pull requests related to open issues -->
<!--- If suggesting a new feature or change, please discuss it in an issue first -->
<!--- If fixing a bug, there should be an issue describing it with steps to reproduce -->
<!--- Please link to the issue here: -->
Closes https://github.com/QuantConnect/Lean/issues/3925
#### Motivation and Context
<!--- Why is this change required? What problem does it solve? -->
Improve performance
#### Requires Documentation Change
<!--- Please indicate if these changes will require updates to documentation, and if so, specify what changes are required -->
N/A
#### How Has This Been Tested?
<!--- Please describe in detail how you tested your changes. -->
<!--- Include details of your testing environment, and the tests you ran to -->
<!--- see how your change affects other areas of the code, etc. -->
Unit, regression and performance tests
#### Types of changes
<!--- What types of changes does your code introduce? Put an `x` in all the boxes that apply: -->
- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] Refactor (non-breaking change which improves implementation)
- [x] Performance (non-breaking change which improves performance. Please add associated performance test and results)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to change)
- [ ] Non-functional change (xml comments/documentation/etc)

#### Checklist:
<!--- The following is a checklist of items that MUST be completed before a PR is accepted -->
<!--- If you're unsure about any of these, don't hesitate to ask. We're here to help! -->
- [x] My code follows the code style of this project.
- [x] I have read the **CONTRIBUTING** [document](https://github.com/QuantConnect/Lean/blob/master/CONTRIBUTING.md).
- [x] I have added tests to cover my changes. <!--- If not applicable, please explain why -->
- [x] All new and existing tests passed.
- [x] My branch follows the naming convention `bug-<issue#>-<description>` or `feature-<issue#>-<description>`

<!--- Template inspired by https://www.talater.com/open-source-templates/#/page/99 -->